Small performance improvements, moved sRGB conversion logic to macro, refactored code for better readability

+++ shaders/include/internal_settings.glsl
@@ -30,7 +30,7 @@
 // SHADOW MAP TEXTURES / SHADOW BUFFER
 // ============================================================================
 
-const int shadowcolor0Format = RGBA16F;             // Shadow color (albedo)
+const int shadowcolor0Format = RGBA8;               // Shadow color (albedo)
 const int shadowcolor1Format = R16F;                // Caustics
 
 const int colortex3Format = RGBA16F;                // Shadow map
